Organização do projeto de acordo com arquitetura MVC-K&P
