+++ Controller/MenuPrincipal.cs
@@ -1,4 +1,7 @@
-﻿
+﻿using GestorGinasio.View.Terminal;
+using GestorGinasio.Model.Entities;
+using GestorGinasio.Model.Services;
+
 namespace GestorGinasio.Controller
 {
     internal class MenuPrincipal
