+++ View/Terminal/GerarRelatoriosView.cs
@@ -1,9 +1,9 @@
 ﻿// File: View/RelatoriosView.cs
 using System;
 using System.Threading;
-using GestorGinasio.Services;
+using GestorGinasio.Model.Services;
 
-namespace GestorGinasio.View
+namespace GestorGinasio.View.Terminal
 {
     public class RelatoriosView
     {
