+++ View/Terminal/README.md
